+++ NEWS.md
@@ -1,5 +1,6 @@
 - Version 0.10.0 (unreleased)
   - Add support for arbitrary clipboard commands - thanks @ign0tus!
+  - `--clipboard` option deprecated in favour of `--clipboard-cmd`
 
 - Version 0.9.0 - released 2020-07-22
   - If `wtype` is available, it will be preferred over `xdotool` when `--type`